+++ Firmware/radio/radio.c
@@ -76,10 +76,12 @@ static void	clear_status_registers(void);
 bool
 radio_receive_packet(uint8_t *length, __xdata uint8_t * __pdata buf)
 {
+#ifdef INCLUDE_GOLAY
 	__xdata uint8_t gout[3];
 	__data uint16_t crc1, crc2;
-	__data uint8_t errcount = 0;
 	__data uint8_t elen;
+#endif // INCLUDE_GOLAY
+	__data uint8_t errcount = 0;
 
 	if (!packet_received) {
 		return false;
@@ -98,14 +100,18 @@ radio_receive_packet(uint8_t *length, __xdata uint8_t * __pdata buf)
 	}
 #endif
 
+// If GOLAY not included always run this code..
+#ifdef INCLUDE_GOLAY
 	if (!feature_golay) {
+#endif
 		// simple unencoded packets
 		*length = receive_packet_length;
 		memcpy(buf, radio_buffer, receive_packet_length);
 		radio_receiver_on();
 		return true;
+		
+#ifdef INCLUDE_GOLAY
 	}
-
 	// decode it in the callers buffer. This relies on the
 	// in-place decode properties of the golay code. Decoding in
 	// this way allows us to overlap decoding with the next receive
@@ -121,7 +127,7 @@ radio_receive_packet(uint8_t *length, __xdata uint8_t * __pdata buf)
 		debug("rx len invalid %u\n", (unsigned)elen);
 		goto failed;
 	}
-
+	
 	// decode the header
 	errcount = golay_decode(6, buf, gout);
 	if (gout[0] != netid[0] ||
@@ -174,7 +180,7 @@ radio_receive_packet(uint8_t *length, __xdata uint8_t * __pdata buf)
 	}
 
 	return true;
-
+#endif // INCLUDE_GOLAY
 failed:
 	if (errors.rx_errors != 0xFFFF) {
 		errors.rx_errors++;
@@ -415,7 +421,7 @@ radio_transmit_simple(__data uint8_t length, __xdata uint8_t * __pdata buf, __pd
 	return false;
 }
 
-
+#ifdef INCLUDE_GOLAY
 // start transmitting a packet from the transmit FIFO
 //
 // @param length		number of data bytes to send
@@ -465,6 +471,7 @@ radio_transmit_golay(uint8_t length, __xdata uint8_t * __pdata buf, __pdata uint
 
 	return radio_transmit_simple(elen, radio_buffer, timeout_ticks);
 }
+#endif // INCLUDE_GOLAY
 
 // start transmitting a packet from the transmit FIFO
 //
@@ -484,11 +491,16 @@ radio_transmit(uint8_t length, __xdata uint8_t * __pdata buf, __pdata uint16_t t
 	PA_ENABLE = 1;		// Set PA_Enable to turn on PA prior to TX cycle
 #endif
 	
+#ifdef INCLUDE_GOLAY
 	if (!feature_golay) {
 		ret = radio_transmit_simple(length, buf, timeout_ticks);
 	} else {
 		ret = radio_transmit_golay(length, buf, timeout_ticks);
 	}
+#else // INCLUDE_GOLAY
+	ret = radio_transmit_simple(length, buf, timeout_ticks);
+#endif // INCLUDE_GOLAY
+	
 #ifdef _BOARD_RFD900A
 	PA_ENABLE = 0;		// Set PA_Enable to off the PA after TX cycle
 #endif