+++ Firmware/radio/radio.h
@@ -58,6 +58,8 @@
 #include "board.h"
 #include "serial.h"
 #include "board_info.h"
+#include "pins_user.h"
+#include "golay23.h"
 #include "parameters.h"
 #include "at.h"
 #include "flash.h"